+++ src/components/Header/Header.js
@@ -2,7 +2,7 @@ import React from "react";
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
 import "./Header.css";
-import logo from "../assets/logo.png";
+import logo from "../../assets/logo.png";
 
 const Header = ({ startDate, endDate, onDateChange }) => {
   return (
